+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## β Beta
 
+#### 2025-03-13 — 0.4.0
+
+-   Added support for Microsoft Edge debugging.
+    -   Requires [Microsoft Edge Tools for VS Code](https://marketplace.visualstudio.com/items?itemName=ms-edgedevtools.vscode-edge-devtools) extension.
+    -   To use, set `attachOptions.action` to `debugWithEdge`.
+
 #### 2025-02-14 — 0.3.0
 
 -   Added support for Mozilla Firefox debugging.